XP-516 Image Focal Point - Fix cropping of images with focal point

+++ modules/core-api/src/main/java/com/enonic/xp/content/Media.java
@@ -85,7 +85,7 @@ public class Media
             return FocalPoint.DEFAULT;
         }
 
-        return new FocalPoint( focalX.floatValue(), focalY.floatValue() );
+        return new FocalPoint( focalX, focalY );
     }
 
     public Attachment getBestFitImageAttachment( int requiredImageSize )
